+++ flaskTodoAPi/models.py
@@ -1,0 +1,16 @@
+from flaskTodoAPi import db, app 
+
+class User(db.Model):
+    id = db.Column(db.Integer,primary_key=True)
+    public_id = db.Column(db.String(50),unique=True)
+    name = db.Column(db.String(50))
+    email = db.Column(db.String(30))
+    password = db.Column(db.String(80))
+    admin = db.Column(db.Boolean)
+
+class Todo(db.Model):
+    id = db.Column(db.Integer,primary_key=True)
+    text = db.Column(db.String(50))
+    complete = db.Column(db.Boolean)
+    user_id = db.Column(db.Integer)   
+
